Intelectual systems [8th semester] /lab1/requirements.txt
nltk==3.5
python-telegram-bot==13.1
